🎯 Streamlined to Rapier2D Only
✨ Major Cleanup:
- Replaced old AABB collision system with Rapier2D as main engine
- Removed gravity (no more falling entities)
- Single, unified codebase with professional physics
- Clean project structure with only Rapier2D dependencies

🎮 User Experience:
- Entities stay in place unless moved by player or collisions
- Perfect collision detection and response
- Smooth physics without gravity interference
- Professional-grade collision handling

🏗️ Architecture:
- main.rs now contains the Rapier2D version
- Removed duplicate binaries and files
- Cleaner, more maintainable codebase
- Ready for future physics features

🎯 Result:
- One engine, maximum performance
- No overlapping collision issues
- Production-ready physics foundation
- Simplified development workflow
